[wip] check option refactoring
