+++ cmd/root.go
@@ -57,7 +57,6 @@ func init() {
 	rootCmd.PersistentFlags().BoolVarP(&strict, "strict", "s", false, "Set strict punycode conversion rules")
 	rootCmd.PersistentFlags().BoolVarP(&fromPuny, "puny", "p", false, "Convert from punycode")
 	rootCmd.PersistentFlags().BoolVarP(&table, "table", "t", false, "Show table of all included unicode characters")
-
 }
 
 func parseFlags(cmd *cobra.Command, args []string) string {
@@ -76,135 +75,6 @@ func parseFlags(cmd *cobra.Command, args []string) string {
 	table, _ := flags.GetBool("table")
 
 	return processInput(args[0], strict, punyDecode, table)
-
-}
-
-// listRanges takes a string and returns a map of Unicode range
-// names and the count of runes within that range
-func listRanges(ustring string) map[string]int {
-	rangeCounts := map[string]int{}
-
-	for _, r := range ustring {
-		rangeCounts[FindRange(r)]++
-	}
-	return rangeCounts
-}
-
-// checkMultipleRange takes a string and determines whether the string
-// contains any characters from outside the predominant range.
-// Runes from the 'Common' range are ignored.
-func checkMultipleRange(ustring string) (multiRange bool) {
-	var ranges int
-	var out string
-	for i, count := range listRanges(ustring) {
-		if i == "Common" {
-			continue
-		}
-		ranges++
-		out += fmt.Sprintf("%s: %d\n", i, count)
-	}
-	if ranges > 1 {
-		multiRange = true
-		fmt.Print(out)
-	}
-	return
-}
-
-// toPuny takes a string and a slice of []idna.Option rules and calls
-// idna.ToASCII, returning the punycode string and error
-func toPuny(s string, rules []idna.Option) (string, error) {
-	punyRules := idna.New(
-		rules...,
-	)
-	return punyRules.ToASCII(s)
-}
-
-// fromPuny takes a punycode string and returns the decoded UTF-8 string
-func fromPuny(s string, rules []idna.Option) (string, error) {
-	punyRules := idna.New(
-		rules...,
-	)
-	return punyRules.ToUnicode(s)
-}
-
-// canPunyConvert takes a rune and a slice of []idna.Option rules and attempts
-// the ToASCII conversion, then returns a bool indicating success
-func canPunyConvert(s string, rules []idna.Option) bool {
-	_, err := toPuny(s, rules)
-	return err == nil
-}
-
-// enumerateErrors takes a rune, checks several punycode conversion rules and
-// reports the failures as a single string
-func enumerateErrors(r rune) []string {
-	rules := map[string][]idna.Option{
-		"CheckBidi (RFC 5893)":                       {idna.BidiRule()},
-		"CheckJoiners (RFC 5892)":                    {idna.CheckJoiners(true)},
-		"CheckHyphens (UTS 46)":                      {idna.CheckHyphens(true)},
-		"ValidateForRegistration (RFC 5891)":         {idna.ValidateForRegistration()},
-		"ValidateLabels (RFC 5891)":                  {idna.ValidateLabels(true)},
-		"UseSTD3ASCIIRules (RFC 1034, 5891, UTS 46)": {idna.StrictDomainName(true), idna.ValidateLabels(true)},
-	}
-
-	var allErrors []string
-
-	for i, ruleset := range rules {
-		if !canPunyConvert(string(r), ruleset) {
-			allErrors = append(allErrors, i)
-		}
-	}
-
-	return allErrors
-
-}
-
-// politePrint takes a rune and outputs a human readable conversion
-func politePrint(r rune) string {
-	// there's a box of mysteries to explore if you print random crap
-	// into a terminal. Let's try to be nice and avoid trashing the
-	// user's environment. It was their input, but they could be
-	// drunk. Or a cat.
-	politeCharmap := []string{
-		"^@", "^A", "^B", "^C", "^D", "^E", "^F", "^G", "^H", "^I", "^J", "^K", "^L",
-		"^M", "^N", "^O", "^P", "^Q", "^R", "^S", "^T", "^U", "^V", "^W", "^X", "^Y",
-		"^Z", "^[", "^\\", "^]", "^^", "^_",
-	}
-
-	switch {
-	case 0x035C <= r && r <= 0x0362: // combining diacritical marks, 2 characters
-		return " ◌" + string(r) + "◌"
-	case 0x0300 <= r && r <= 0x036F: // combining diacritical marks
-		return "  ◌" + string(r)
-	// filter control characters
-	// todo: replace these filters with ranges defined in unicode/table.go
-	// reasoning: terminal control sequences can do all sorts of damage to the
-	// output.  we will remove them and put in the caret notation for C0 and 'C1'
-	// for C1 unicode control characters
-	case int(r) < len(politeCharmap):
-		return politeCharmap[r] // C0 controls
-	case int(r) == 127:
-		return "^?" // DEL
-	case 0x206A <= r && r <= 0x206F: // deprecated format characters
-		return " "
-	case 0x80 <= r && r <= 0x9F:
-		return "C1" // Unicode C1 controls
-	// filter direction changing characters
-	// reasoning: we are printing a single character here. If we needed to print
-	// words then the directional marks should not be discarded. These checks are
-	// an attempt to prevent leaving an unclosed direction change in the output.
-	// ref. https://www.unicode.org/reports/tr9/#Directional_Formatting_Codes
-	case 0x202A <= r && r <= 0x202E: // directional overrides (LRE, RLE, PDF, LRO, RLO)
-		return " "
-	case r == 0x061C, r == 0x200E, r == 0x200F: // implicit directional marks
-		return " "
-	case 0x2066 <= r && r <= 0x2069: // directional isolates (LRI, RLI, FSI, PDI)
-		return " "
-	case r == 0x200d || r == 0x2060: // joiners
-		return " "
-	case 0xFE00 <= r && r <= 0xFE0F: // variation selectors
-		return " "
-	}
-	return string(r)
 }
 
 // toString takes a rune returns a string with padding appropriate for the character width
